requirements.txt
requests==2.31.0
supabase==1.0.4
python-dotenv==1.0.0
flask==2.3.3
beautifulsoup4==4.12.2
